+++ examples/05-scale-showcase/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @cotal-ai/example-05-scale-showcase
 
+## 0.0.7
+
+### Patch Changes
+
+- @cotal-ai/core@0.33.1
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [ba74c84]
+  - @cotal-ai/core@0.33.0
+
+## 0.0.5
+
+### Patch Changes
+
+- @cotal-ai/core@0.32.0
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [4ef59c3]
+  - @cotal-ai/core@0.31.0
+
 ## 0.0.3
 
 ### Patch Changes